put tree models in their own module
